# api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Les Canetons API",
        "version": "1.0.0",
...
    },
    "servers": [
        {
            "url": "https://lescanetons.org/api/v1",
            "description": "Production"
        }
    ],
...
    "paths": {
...
        "/events/{event}/registration-options": {
            "put": {
                "operationId": "registrationOption.replace",
                "description": "Requires `events.manage`, because configuring what an event offers is\nconfiguring the event.\n\nSend the complete list. An entry carrying an `id` updates that option,\nan entry without one creates it, and an option missing from the list\nis deleted. Replaying the same body converges on the same list, so a\nretry after a timeout is safe.\n\n`priceCents` is a whole number of centimes and may be null, for an\noption that is free or whose price lives in its description.\n`sortOrder` may be left out, in which case the entry's position in the\nlist is used.\n\nReturns the event's options as they now stand, in sort order.\n\nRemoving an option somebody has already booked answers\n`409 option_has_registrations`, names the options in the way, and\nchanges nothing. An `id` belonging to another event's option, more\nthan 50 entries, or a missing label answers `400 validation_failed`.",
                "summary": "Replace an event's bookable options",
                "tags": [
                    "Registration"
                ],
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "required": true,
                        "description": "The event ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReplaceRegistrationOptionsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Array of `RegistrationOptionResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RegistrationOptionResource"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "integer",
                                            "const": 409
                                        },
                                        "code": {
                                            "type": "string",
                                            "const": "option_has_registrations"
                                        },
                                        "instance": {
                                            "type": "string"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "requestId": {
                                            "type": "string"
                                        },
                                        "detail": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "title",
                                        "status",
                                        "code",
                                        "instance",
                                        "errors",
                                        "requestId",
                                        "detail"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            }
        },
...
        "/roles": {
            "get": {
                "operationId": "role.index",
                "description": "Requires `members.manage`. Read-only reference data, for a client\nbuilding the roster form.\n\nReturns every role with the permissions it grants. A role carries no\ndisplay name: `key` is the fixed identifier and the client translates\nit.",
                "summary": "List the roles",
                "tags": [
                    "Members"
                ],
                "responses": {
                    "200": {
                        "description": "Array of `RoleResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/RoleResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    }
                }
            }
        },
        "/members": {
            "get": {
                "operationId": "member.index",
                "description": "Requires `members.manage`. Returns everyone the band tracks, ordered by\nlast name then first name. Each entry carries the person's register\n(`sectionId` and `sectionName`), whether they play (`isPlayer`), their\ncommittee title, whether they may be shown on the public site, and\n`roleIds`, the roles they hold.\n\nNo password and no hash is ever included, and neither are effective\npermissions: a role is what grants them, so read `GET /api/v1/roles` and\njoin on `roleIds`.",
                "summary": "List the roster",
                "tags": [
                    "Members"
                ],
                "responses": {
                    "200": {
                        "description": "Array of `MemberResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MemberResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    }
                }
            },
            "post": {
                "operationId": "member.store",
                "description": "Requires `members.manage`. Creating a person creates their account:\nevery member has one. Send `firstName`, `lastName`, `username` and\n`publicVisible`, and optionally `sectionId`, `committeeTitle` and\n`instructorOfSectionId`.\n\nAnswers `201` with `member` and `generatedPassword`, a readable password\nthe server minted for the new account. It is shown in this response and\nnowhere else: it is hashed on the way into the database, is never\nwritten to the audit log, and no later call returns it. An administrator\nwho loses it issues a new one at\n`POST /api/v1/members/{member}/password`. The new member is required to\nchange it before doing anything else.\n\nNo roles are granted. Roles are `PUT /api/v1/members/{member}/roles`, and\nno password may be chosen here.\n\nA missing required field answers `400 validation_failed` naming the\nfield with `required`. A username already in use answers the same with\n`already_taken`; one containing anything but lower-case letters, digits,\ndot, hyphen or underscore answers `invalid_format`.",
                "summary": "Add a member to the roster",
                "tags": [
                    "Members"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMemberRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "member": {
                                            "$ref": "#/components/schemas/MemberResource"
                                        },
                                        "generatedPassword": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "member",
                                        "generatedPassword"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            }
        },
...
        "/members/{member}/roles": {
            "put": {
                "operationId": "memberRole.replace",
                "description": "Requires `members.manage`. `roleIds` is the complete set the member is\nleft with, so send every role they should keep; an empty array removes\nall of them. Read the roles that exist from `GET /api/v1/roles`.\n\nThis is the only way any permission is granted or taken away. A role is\nwhat groups permissions, and no endpoint in this API authorises by role\nname, so nothing is ever granted to one member directly.\n\nThe change ends every session the member has open, so a permission that\nhas just been removed stops working immediately rather than at their\nnext login. Answers the updated member and `sessionsEnded`.\n\nRefuses with `409 cannot_remove_last_administrator` when the new set\nwould leave nobody able to administer members, and\n`409 cannot_demote_self` when callers would strip their own member\nadministration. Both are `409` rather than `403` because the caller does\nhold the permission; the request conflicts with the state of the roster.\nWhen both apply, the last-administrator refusal is the one returned. An\nunknown role id answers `400 validation_failed` against `roleIds`.",
                "summary": "Replace a member's roles",
                "tags": [
                    "Members"
                ],
                "parameters": [
                    {
                        "name": "member",
                        "in": "path",
                        "required": true,
                        "description": "The member ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReplaceMemberRolesRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "member": {
                                            "$ref": "#/components/schemas/MemberResource"
                                        },
                                        "sessionsEnded": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "member",
                                        "sessionsEnded"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "400": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "403": {
                        "$ref": "#/components/responses/Problem403"
                    },
                    "419": {
                        "$ref": "#/components/responses/Problem419"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RegistrationOptionResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "label": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "priceCents": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "description": "Price in centimes: 4500 is CHF 45.00. Null when the option is free or priced in its description."
                    },
                    "sortOrder": {
                        "type": "integer",
                        "description": "Ascending display order."
                    }
                },
                "required": [
                    "id",
                    "label",
                    "description",
                    "priceCents",
                    "sortOrder"
                ],
                "title": "RegistrationOptionResource"
            },
...
            "ReplaceMemberRolesRequest": {
                "type": "object",
                "description": "The complete set of roles a member is to be left with.\n\nThis is a replacement, not a partial update: send every role the member\nshould hold afterwards, including the ones they already have. Anything left\nout is removed, and an empty array removes all of them.\n\n`roleIds` must be sent, even when empty. Read the ids that exist from\n`GET /api/v1/roles`; an unknown one fails validation against `roleIds`.",
                "properties": {
                    "roleIds": {
                        "type": "array",
                        "description": "Every role the member should hold afterwards. Send `[]` to leave them with none.",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "required": [
                    "roleIds"
                ],
                "title": "ReplaceMemberRolesRequest"
            },
            "ReplaceRegistrationOptionsRequest": {
                "type": "object",
                "description": "The complete list of things that can be booked at one event.\n\nThis is a replacement, not a partial update. Send the whole list every\ntime: an entry carrying an `id` updates that option, an entry without one\ncreates a new option, and any existing option missing from the list is\ndeleted. Deleting an option somebody has already booked is refused with\n`option_has_registrations` rather than rewriting what that person ordered.\n\nSend `options: []` to leave the event with nothing bookable. At most 50\nentries. An `id` must belong to this event.\n\n`priceCents` is an integer number of centimes, so 4500 is CHF 45.00. Zero\nis a free option; `null` means the option carries no price at all, for\ninstance when the price is spelled out in the description.",
                "properties": {
                    "options": {
                        "type": "array",
                        "description": "The whole list, replacing whatever the event offered before. Send `[]` to offer nothing.",
                        "examples": [
                            [
                                {
                                    "id": 7,
                                    "label": "Repas adulte",
                                    "description": "Jambon, gratin et salade",
                                    "priceCents": 4500,
                                    "sortOrder": 0
                                },
                                {
                                    "label": "Repas enfant",
                                    "description": null,
                                    "priceCents": 2000,
                                    "sortOrder": 1
                                }
                            ]
                        ],
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": [
                                        "integer",
                                        "null"
                                    ],
                                    "description": "The id of an existing option of this event, to update it. Omit or send `null` to create a new one."
                                },
                                "label": {
                                    "type": "string",
                                    "description": "What the guest sees on the booking form, for example \"Repas adulte\".",
                                    "maxLength": 255
                                },
                                "description": {
                                    "type": [
                                        "string",
                                        "null"
                                    ],
                                    "maxLength": 255
                                },
                                "priceCents": {
                                    "type": [
                                        "integer",
                                        "null"
                                    ],
                                    "description": "Integer centimes: `4500` is CHF 45.00. `0` is free; `null` is an option with no price of its own.",
                                    "minimum": 0,
                                    "maximum": 1000000
                                },
                                "sortOrder": {
                                    "type": [
                                        "integer",
                                        "null"
                                    ],
                                    "description": "Where this option sits on the form, lowest first, ties broken by id. Omitted, it takes the entry position in the list you send.",
                                    "minimum": 0,
                                    "maximum": 1000
                                }
                            },
                            "required": [
                                "label"
                            ]
                        },
                        "maxItems": 50
                    }
                },
                "required": [
                    "options"
                ],
                "title": "ReplaceRegistrationOptionsRequest"
            },
            "RoleResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "key": {
                        "type": "string",
                        "description": "Stable identifier. Roles carry no display name; clients translate by this key."
                    },
                    "permissions": {
                        "type": "array",
                        "description": "Everything this role grants, as permission strings.",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "id",
                    "key",
                    "permissions"
                ],
                "title": "RoleResource"
            },
...
        },
        "responses": {
...
            "ValidationException": {
                "description": "The submitted fields were rejected. `errors` names each one and why.",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "description": "English message. Never displayed: the front end renders `code`."
                                },
                                "status": {
                                    "type": "integer",
                                    "description": "The HTTP status code, repeated in the body as RFC 9457 intends."
                                },
                                "instance": {
                                    "type": "string",
                                    "description": "The path that was requested. Never the query string."
                                },
                                "code": {
                                    "type": "string",
                                    "description": "Stable machine token the front end maps to French.",
                                    "enum": [
                                        "validation_failed"
                                    ]
                                },
                                "errors": {
                                    "type": "array",
                                    "description": "One entry per rejected field, first failure only.",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "description": "The camelCase name of the rejected field, as it was submitted."
                                            },
                                            "reason": {
                                                "type": "string",
                                                "description": "Stable machine token saying what is wrong with that field.",
                                                "enum": [
                                                    "required",
                                                    "too_long",
                                                    "too_short",
                                                    "invalid_format",
                                                    "invalid_type",
                                                    "invalid_value",
                                                    "invalid_number",
                                                    "already_taken",
                                                    "must_be_after"
                                                ]
                                            },
                                            "params": {
                                                "type": "object",
                                                "additionalProperties": {}
                                            }
                                        },
                                        "required": [
                                            "field",
                                            "reason"
                                        ]
                                    }
                                },
                                "requestId": {
                                    "type": "string",
                                    "description": "ULID identifying this request. Echoed as the X-Request-Id header, and in the logs."
                                },
                                "detail": {
                                    "type": "string",
                                    "description": "RFC 9457 \u00a73.1.1: what happened and what to do about it, in English, for a developer reading the response. Never rendered to an end user."
                                }
                            },
                            "required": [
                                "title",
                                "status",
                                "code",
                                "instance",
                                "errors",
                                "requestId",
                                "detail"
                            ]
                        }
                    }
                }
            },
...
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "No session, or it has expired. Log in and retry.",
                "content": {
                    "application/problem+json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "description": "English message. Never displayed: the front end renders `code`."
                                },
                                "status": {
                                    "type": "integer",
                                    "description": "The HTTP status code, repeated in the body as RFC 9457 intends."
                                },
                                "instance": {
                                    "type": "string",
                                    "description": "The path that was requested. Never the query string."
                                },
                                "code": {
                                    "type": "string",
                                    "description": "Stable machine token the front end maps to French.",
                                    "enum": [
                                        "not_authenticated"
                                    ]
                                },
                                "errors": {
                                    "type": "array",
                                    "description": "Always empty for this status; present so every problem has the same shape.",
                                    "items": {
                                        "type": "object"
                                    }
                                },
                                "requestId": {
                                    "type": "string",
                                    "description": "ULID identifying this request. Echoed as the X-Request-Id header, and in the logs."
                                },
                                "detail": {
                                    "type": "string",
                                    "description": "RFC 9457 \u00a73.1.1: what happened and what to do about it, in English, for a developer reading the response. Never rendered to an end user."
                                }
                            },
                            "required": [
                                "title",
                                "status",
                                "code",
                                "instance",
                                "errors",
                                "requestId",
                                "detail"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}